jobdoor/requirements.txt
asgiref==3.2.10
certifi==2022.12.7
charset-normalizer==3.1.0
defusedxml==0.7.1
Django==3.1.2
django-allauth==0.42.0
django-filter==23.1
django-rest-auth==0.9.5
djangorestframework==3.12.1
idna==3.4
oauthlib==3.2.2
Pillow==9.5.0
python3-openid==3.2.0
pytz==2023.3
requests==2.28.2
requests-oauthlib==1.3.1
six==1.16.0
sqlparse==0.4.3
tzdata==2023.3
urllib3==1.26.15
